refactor(bundles): extract IdentityManager/MemoryManager components
PR 2 of SPEC_BUNDLE_MANAGEMENT_2026_05_22.md. Pure refactor — zero
behavior or UX change. Extracts the Identity and Memory pane bodies
into context-free components so the same UI can render both in the
existing agent-settings tab and (in a later PR) inside a window modal.

- New memory-manager.tsx: MemoryManagerBody (the rail+detail UI driven
  by a MemoryViewModel) + MemoryManager (context-free, owns its own
  block-free model). memory-view.tsx is now a thin wrapper rendering
  MemoryManagerBody with the pane's BlockRegistry model.
- New identity-manager.tsx: IdentityManagerBody + IdentityManager,
  same shape. identity-pane-view.tsx is now a thin wrapper.
- The pane bodies never touched blockId/nodeModel directly — only
  model.* accessors. The only block coupling was in the ViewModels:
  blockAtom + the cosmetic viewName header memo. Made both ViewModel
  constructors accept (blockId?, nodeModel?); blockAtom resolves to a
  () => undefined accessor when no block is supplied, so the manager
  can construct a model with no Agent pane behind it.
- BlockRegistry still calls `new ctor(blockId, nodeModel)` for the
  `view: "identity"` / `view: "memory"` panes, unchanged — optional
  params accept the existing required-arg call site.

The settings tabs render the extracted body verbatim (same JSX, same
class names, same SCSS import), so they are byte-for-byte unchanged.
