+++ .changesets/1779443302-refactor-bundles-extract-identitymanager-memorymanager-compo-bpcv.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(bundles): extract IdentityManager/MemoryManager components
